Davide/travel/requirements.txt
flask
flask-restful
requests
wtforms
google-cloud-firestore
gunicorn
marshmallow